refactor: move insight period glance types
